Split incident updates out of incident controller
